[feat] SSE Response 구현

+++ SMASHING/Networks/Response/SSE/SSEMatchingReceivedPayload.swift
@@ -10,7 +10,7 @@ import Foundation
 struct SSEMatchingReceivedPayload: Codable {
     let type: String
     let matchingId: String
-    let sportId: Int64
+    let sportId: Int
     let receiverProfileId: String
     let requester: SSERequesterSummary
 }
